+++ aimmo-game/tests/test_simulation/test_world_map.py
@@ -7,7 +7,7 @@ from unittest import TestCase
 
 from simulation.location import Location
 from simulation.world_map import WorldMap, world_map_static_spawn_decorator
-from simulation.maps.cell import Cell
+from simulation.geography.cell import Cell
 from .dummy_avatar import DummyAvatar
 from .maps import MockCell, MockPickup
